Code Organization Changes, Improving class encapsulation - modelfitting.py
