--- whobpyt/models/jansen_rit.py
@@ -1,143 +1,0 @@
-"""
-Authors: Zheng Wang, John Griffiths, Andrew Clappison, Hussain Ather
-Neural Mass Model fitting
-module for JR with forward backward and lateral connection for EEG
-"""
-
-# @title new function PyTepFit
-
-# Pytorch stuff
-
-
-"""
-Importage
-"""
-import torch
-from whobpyt.datatypes.AbstractParams import AbstractParams
-from whobpyt.datatypes.AbstractNMM import AbstractNMM
-from whobpyt.functions.pytorch_funs import setModelParameters
-from whobpyt.functions.pytorch_funs import integration_forward
-
-
-class ParamsJR(AbstractParams):
-
-    def __init__(self, **kwargs):
-
-        param = {
-            "A ": [3.25, 0], 
-            "a": [100, 0.], 
-            "B": [22, 0], 
-            "b": [50, 0], 
-            "g": [1000, 0],
-            
-            "c1": [135, 0.], 
-            "c2": [135 * 0.8, 0.], 
-            "c3 ": [135 * 0.25, 0.], 
-            "c4": [135 * 0.25, 0.],
-            
-            "std_in": [100, 0], 
-            "vmax": [5, 0], 
-            "v0": [6, 0], 
-            "r": [0.56, 0], 
-            "y0": [2, 0],
-            
-            "mu": [.5, 0], 
-            "k": [5, 0], 
-            "cy0": [5, 0], 
-            "ki": [1, 0]
-        }
-        
-        for var in param:
-            setattr(self, var, param[var])
-
-        for var in kwargs:
-            setattr(self, var, kwargs[var])
-
-
-class RNNJANSEN(AbstractNMM):
-    """
-    A module for forward model (JansenRit) to simulate a batch of EEG signals
-    Attibutes
-    ---------
-    state_size : int
-        the number of states in the JansenRit model
-    input_size : int
-        the number of states with noise as input
-    tr : float
-        tr of image
-    step_size: float
-        Integration step for forward model
-    hidden_size: int
-        the number of step_size in a tr
-    TRs_per_window: int
-        the number of EEG signals to simulate
-    node_size: int
-        the number of ROIs
-    sc: float node_size x node_size array
-        structural connectivity
-    fit_gains: bool
-        flag for fitting gains 1: fit 0: not fit
-    g, c1, c2, c3,c4: tensor with gradient on
-        model parameters to be fit
-    w_bb: tensor with node_size x node_size (grad on depends on fit_gains)
-        connection gains
-    std_in std_out: tensor with gradient on
-        std for state noise and output noise
-    hyper parameters for prior distribution of model parameters
-    Methods
-    -------
-    forward(input, noise_out, hx)
-        forward model (JansenRit) for generating a number of EEG signals with current model parameters
-    """
-    state_names = ['E', 'Ev', 'I', 'Iv', 'P', 'Pv']
-    model_name = "JR"
-
-    def __init__(self, node_size: int,
-                 TRs_per_window: int, step_size: float, output_size: int, tr: float, sc: float, lm: float, dist: float,
-                 use_fit_gains: bool, use_fit_lfm: bool, param: ParamsJR) -> None:
-        """
-        Parameters
-        ----------
-
-        tr : float
-            tr of image
-        step_size: float
-            Integration step for forward model
-        
-        TRs_per_window: int
-            the number of EEG signals to simulate
-        node_size: int
-            the number of ROIs
-        output_size: int
-            the number of channels EEG
-        sc: float node_size x node_size array
-            structural connectivity
-        use_fit_gains: bool
-            flag for fitting gains 1: fit 0: not fit
-        use_fit_lfm: bool
-            flag for fitting gains 1: fit 0: not fit
-        param from ParamJR
-        """
-        super(RNNJANSEN, self).__init__()
-        self.state_size = 6  # 6 states JR model
-        self.tr = tr  # tr ms (integration step 0.1 ms)
-        self.step_size = torch.tensor(step_size, dtype=torch.float32)  # integration step 0.1 ms
-        self.steps_per_TR = int(tr / step_size)
-        self.TRs_per_window = TRs_per_window  # size of the batch used at each step
-        self.node_size = node_size  # num of ROI
-        self.output_size = output_size  # num of EEG channels
-        self.sc = sc  # matrix node_size x node_size structure connectivity
-        self.dist = torch.tensor(dist, dtype=torch.float32)
-        self.lm = lm
-        self.use_fit_gains = use_fit_gains  # flag for fitting gains
-        self.use_fit_lfm = use_fit_lfm
-        self.param = param
-
-        self.output_size = lm.shape[0]  # number of EEG channels
-
-    def setModelParameters(self):
-        # set states E I f v mean and 1/sqrt(variance)
-        return setModelParameters(self)
-
-    def forward(self, external, hx, hE):
-        return integration_forward(self, external, hx, hE)
